Sql Agent/requirements.txt
streamlit
pandas
duckdb
google-generativeai
matplotlib
numpy
seaborn 